+++ negotiation/src/main/java/it/eng/negotiation/service/ContractNegotiationEventHandlerService.java
@@ -1,17 +1,23 @@
 package it.eng.negotiation.service;
 
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
 import it.eng.negotiation.exception.ContractNegotiationAPIException;
+import it.eng.negotiation.exception.ContractNegotiationNotFoundException;
 import it.eng.negotiation.model.Agreement;
 import it.eng.negotiation.model.ContractAgreementMessage;
 import it.eng.negotiation.model.ContractAgreementVerificationMessage;
 import it.eng.negotiation.model.ContractNegotiation;
 import it.eng.negotiation.model.ContractNegotiationState;
+import it.eng.negotiation.model.ContractNegotiationTerminationMessage;
 import it.eng.negotiation.model.Offer;
+import it.eng.negotiation.model.Reason;
 import it.eng.negotiation.properties.ContractNegotiationProperties;
 import it.eng.negotiation.repository.AgreementRepository;
 import it.eng.negotiation.repository.ContractNegotiationRepository;
@@ -20,23 +26,29 @@ import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
 import it.eng.tools.event.contractnegotiation.ContractNegotiationOfferResponseEvent;
 import it.eng.tools.response.GenericApiResponse;
+import it.eng.tools.util.CredentialUtils;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Slf4j
 public class ContractNegotiationEventHandlerService {
 	
+	private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
+	
 	private final ContractNegotiationRepository contractNegotiationRepository;
 	private final AgreementRepository agreementRepository;
 	private final ContractNegotiationProperties properties;
 	private final OkHttpRestClient okHttpRestClient;
+	private final CredentialUtils credentialUtils;
 	
 	public ContractNegotiationEventHandlerService(ContractNegotiationRepository contractNegotiationRepository,
-			ContractNegotiationProperties properties, OkHttpRestClient okHttpRestClient, AgreementRepository agreementRepository) {
+			ContractNegotiationProperties properties, OkHttpRestClient okHttpRestClient,
+			AgreementRepository agreementRepository, CredentialUtils credentialUtils) {
 		this.contractNegotiationRepository = contractNegotiationRepository;
 		this.agreementRepository = agreementRepository;
 		this.properties = properties;
 		this.okHttpRestClient = okHttpRestClient;
+		this.credentialUtils = credentialUtils;
 	}
 
 	public void handleContractNegotiationOfferResponse(ContractNegotiationOfferResponseEvent offerResponse) {
@@ -52,14 +64,13 @@ public class ContractNegotiationEventHandlerService {
 					.consumerPid(contractNegotiation.getConsumerPid())
 					.providerPid(contractNegotiation.getProviderPid())
 					.callbackAddress(properties.providerCallbackAddress())
-					.agreement(agreementFromOffer(Serializer.deserializePlain(offerResponse.getOffer().toPrettyString(), Offer.class)))
+					.agreement(agreementFromOffer(Serializer.deserializePlain(offerResponse.getOffer().toPrettyString(), Offer.class), contractNegotiation.getAssigner()))
 					.build();
 			
-			String authorization = okhttp3.Credentials.basic("connector@mail.com", "password");
 			GenericApiResponse<String> response = okHttpRestClient.sendRequestProtocol(
-					contractNegotiation.getCallbackAddress() + ContractNegotiationCallback.getContractAgreementCallback("consumer", contractNegotiation.getConsumerPid()), 
+					ContractNegotiationCallback.getContractAgreementCallback(contractNegotiation.getCallbackAddress(), contractNegotiation.getConsumerPid()), 
 					Serializer.serializeProtocolJsonNode(agreementMessage),
-					authorization);
+					credentialUtils.getConnectorCredentials());
 			if(response.isSuccess()) {
 				log.info("Updating status for negotiation {} to agreed", contractNegotiation.getId());
 				ContractNegotiation contractNegtiationUpdate = ContractNegotiation.Builder.newInstance()
@@ -79,14 +90,42 @@ public class ContractNegotiationEventHandlerService {
 			}
 		}
 	}
+	
+	public ContractNegotiation handleContractNegotiationTerminated(String contractNegotiationId) {
+		ContractNegotiation contractNegotiation = contractNegotiationRepository.findById(contractNegotiationId)
+    		.orElseThrow(() ->
+    			new ContractNegotiationNotFoundException("Contract negotiation with id " + contractNegotiationId + " not found"));
+		// for now just log it; maybe we can publish event?
+		log.info("Contract negotiation with consumerPid {} and providerPid {} declined", contractNegotiation.getConsumerPid(), contractNegotiation.getProviderPid());
+		ContractNegotiationTerminationMessage negotiationTerminatedEventMessage = ContractNegotiationTerminationMessage.Builder.newInstance()
+			.consumerPid(contractNegotiation.getConsumerPid())
+			.providerPid(contractNegotiation.getProviderPid())
+			.code(contractNegotiationId)
+			.reason(Arrays.asList(Reason.Builder.newInstance().language("en").value("Contract negotiation terminated by provider").build()))
+			.build();
+			
+		GenericApiResponse<String> response = okHttpRestClient.sendRequestProtocol(
+				ContractNegotiationCallback.getContractTerminationCallback(contractNegotiation.getCallbackAddress(), contractNegotiation.getConsumerPid()), 
+				Serializer.serializeProtocolJsonNode(negotiationTerminatedEventMessage),
+				credentialUtils.getConnectorCredentials());
+		if(response.isSuccess()) {
+			log.info("Updating status for negotiation {} to terminated", contractNegotiation.getId());
+			ContractNegotiation contractNegtiationTerminated = contractNegotiation.withNewContractNegotiationState(ContractNegotiationState.TERMINATED);
+			contractNegotiationRepository.save(contractNegtiationTerminated);
+			return contractNegtiationTerminated;
+		} else {
+			log.error("Response status not 200 - consumer did not process AgreementMessage correct");
+			throw new ContractNegotiationAPIException("consumer did not process AgreementMessage correct");
+		}
+	}
 
-	private Agreement agreementFromOffer(Offer offer) {
+	private Agreement agreementFromOffer(Offer offer, String assigner) {
 		return Agreement.Builder.newInstance()
 				.id(UUID.randomUUID().toString())
-				.assignee(properties.providerCallbackAddress())
-				.assigner(properties.providerCallbackAddress())
-				.target("target")
-				.timestamp("now")
+				.assignee(properties.getAssignee())
+				.assigner(assigner)
+				.target(offer.getTarget())
+				.timestamp(FORMATTER.format(ZonedDateTime.now()))
 				.permission(offer.getPermission())
 				.build();
 	}
@@ -104,14 +143,13 @@ public class ContractNegotiationEventHandlerService {
 		}
 		log.info("Found intial negotiation" + " - CallbackAddress " + contractNegotiation.getCallbackAddress());
 
-		String authorization = okhttp3.Credentials.basic("connector@mail.com", "password");
 		String callbackAddress = ContractNegotiationCallback.getProviderAgreementVerificationCallback(contractNegotiation.getCallbackAddress(), verificationMessage.getProviderPid());
 		log.info("Sending verification message to provider to {}", callbackAddress);
 		GenericApiResponse<String> response = okHttpRestClient.sendRequestProtocol(callbackAddress, 
 				Serializer.serializeProtocolJsonNode(verificationMessage),
-				authorization);
+				credentialUtils.getConnectorCredentials());
 		
-		if(response.getHttpStatus() == 200) {
+		if(response.isSuccess()) {
 			log.info("Updating status for negotiation {} to verified", contractNegotiation.getId());
 			ContractNegotiation contractNegtiationUpdate = ContractNegotiation.Builder.newInstance()
 					.id(contractNegotiation.getId())