+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationAPIServiceTest.java
@@ -1,5 +1,7 @@
 package it.eng.negotiation.service;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -7,7 +9,10 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Optional;
+import java.util.UUID;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,9 +26,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.negotiation.exception.ContractNegotiationAPIException;
+import it.eng.negotiation.exception.OfferNotFoundException;
 import it.eng.negotiation.model.Agreement;
 import it.eng.negotiation.model.ContractNegotiation;
+import it.eng.negotiation.model.ContractNegotiationState;
 import it.eng.negotiation.model.ModelUtil;
+import it.eng.negotiation.model.Offer;
 import it.eng.negotiation.properties.ContractNegotiationProperties;
 import it.eng.negotiation.repository.AgreementRepository;
 import it.eng.negotiation.repository.ContractNegotiationRepository;
@@ -31,6 +39,7 @@ import it.eng.negotiation.repository.OfferRepository;
 import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
 import it.eng.tools.response.GenericApiResponse;
+import it.eng.tools.util.CredentialUtils;
 
 @ExtendWith(MockitoExtension.class)
 public class ContractNegotiationAPIServiceTest {
@@ -47,9 +56,13 @@ public class ContractNegotiationAPIServiceTest {
 	private ContractNegotiationProperties properties;
 	@Mock
 	private GenericApiResponse<String> apiResponse;
+	@Mock
+    private CredentialUtils credentialUtils;
 	
 	@Captor
 	private ArgumentCaptor<ContractNegotiation> argumentCaptor;
+	@Captor
+	private ArgumentCaptor<Agreement> argCaptorAgreement;
 
 	@InjectMocks
 	private ContractNegotiationAPIService service;
@@ -57,21 +70,23 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Start contract negotiation success")
 	public void startNegotiation_success() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.getData()).thenReturn(Serializer.serializeProtocol(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
-		when(apiResponse.getHttpStatus()).thenReturn(201);
+		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.consumerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		
 		service.startNegotiation(ModelUtil.FORWARD_TO, Serializer.serializePlainJsonNode(ModelUtil.OFFER));
 		
 		verify(contractNegotiationRepository).save(any(ContractNegotiation.class));
+		verify(offerRepository).save(any(Offer.class));
 	}
 	
 	@Test
 	@DisplayName("Start contract negotiation failed")
 	public void startNegotiation_failed() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(400);
 		when(properties.consumerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		
 		assertThrows(ContractNegotiationAPIException.class, ()-> service.startNegotiation(ModelUtil.FORWARD_TO, Serializer.serializePlainJsonNode(ModelUtil.OFFER)));
@@ -82,9 +97,10 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Start contract negotiation json exception")
 	public void startNegotiation_jsonException() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.getData()).thenReturn("not a JSON");
-		when(apiResponse.getHttpStatus()).thenReturn(201);
+		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.consumerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		
 		assertThrows(ContractNegotiationAPIException.class, ()-> service.startNegotiation(ModelUtil.FORWARD_TO, Serializer.serializePlainJsonNode(ModelUtil.OFFER)));
@@ -96,6 +112,7 @@ public class ContractNegotiationAPIServiceTest {
 	@DisplayName("Process posted offer - success")
 	public void postContractOffer_success() {
 		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class)))
 			.thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(true);
@@ -109,6 +126,7 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Process posted offer - error")
 	public void postContractOffer_error() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class)))
 			.thenReturn(apiResponse);
@@ -121,8 +139,9 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Send agreement success - accepted state")
 	public void sendAgreement_success_acceptedState() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(200);
+		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString())).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
 		
@@ -135,8 +154,9 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Send agreement success - requested state")
 	public void sendAgreement_success_requestedState() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(200);
+		when(apiResponse.isSuccess()).thenReturn(true);
 		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString())).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
 		
@@ -172,8 +192,8 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Send agreement failed - bad request")
 	public void sendAgreement_failedBadRequest() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(400);
 		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
 		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString())).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
 		
@@ -187,6 +207,7 @@ public class ContractNegotiationAPIServiceTest {
 	@Test
 	@DisplayName("Finalize negotiation success")
 	public void finalizeNegotiation_success_requestedState() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(true);
 		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString())).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_VERIFIED));
@@ -212,7 +233,8 @@ public class ContractNegotiationAPIServiceTest {
 		
 		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(ModelUtil.PROVIDER_PID, ModelUtil.CONSUMER_PID)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_OFFERED));
 
-		assertThrows(ContractNegotiationAPIException.class, () -> service.finalizeNegotiation(ModelUtil.CONTRACT_NEGOTIATION_EVENT_MESSAGE));
+		assertThrows(ContractNegotiationAPIException.class, 
+				() -> service.finalizeNegotiation(ModelUtil.CONTRACT_NEGOTIATION_EVENT_MESSAGE));
 		
 		verify(contractNegotiationRepository).findByProviderPidAndConsumerPid(ModelUtil.PROVIDER_PID, ModelUtil.CONSUMER_PID);
 		verify(contractNegotiationRepository, times(0)).save(any(ContractNegotiation.class));
@@ -220,15 +242,150 @@ public class ContractNegotiationAPIServiceTest {
 	}
 	
 	@Test
+	@DisplayName("Finalize negotiation error - already finalized")
+	public void finalizeNegotiation_error_finalized_state() {
+		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString()))
+			.thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_FINALIZED));
+		
+		assertThrows(ContractNegotiationAPIException.class,
+				() -> service.finalizeNegotiation(ModelUtil.CONTRACT_NEGOTIATION_EVENT_MESSAGE));
+		
+		verify(contractNegotiationRepository, times(0)).save(argumentCaptor.capture());
+	}
+	
+	@Test
 	@DisplayName("Finalize negotiation failed - bad request")
 	public void finalizeNegotiation_failedBadRequest() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(false);
-		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString())).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_VERIFIED));
+		when(contractNegotiationRepository.findByProviderPidAndConsumerPid(anyString(), anyString()))
+			.thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_VERIFIED));
 		
 		assertThrows(ContractNegotiationAPIException.class, ()-> service.finalizeNegotiation(ModelUtil.CONTRACT_NEGOTIATION_EVENT_MESSAGE));
 	
 		verify(okHttpRestClient).sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class));
 		verify(contractNegotiationRepository, times(0)).save(any(ContractNegotiation.class));
 	}
+	
+	@Test
+	@DisplayName("Find all contract negotiations")
+	public void findAll() {
+		when(contractNegotiationRepository.findAll())
+				.thenReturn(Arrays.asList(
+						ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED,
+						ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		Collection<JsonNode> response = service.findContractNegotiations(null);
+		assertNotNull(response);
+		assertEquals(2, response.size());
+	}
+	
+	@Test
+	@DisplayName("Find contract negotiations by state")
+	public void findContractNegotiationByState() {
+		when(contractNegotiationRepository.findByState(ContractNegotiationState.ACCEPTED.name()))
+				.thenReturn(Arrays.asList(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
+		Collection<JsonNode> response = service.findContractNegotiations(ContractNegotiationState.ACCEPTED.name());
+		assertNotNull(response);
+		assertEquals(1, response.size());
+	}
+	
+	@Test
+	@DisplayName("Consumer accepts contract negotiation offered by provider")
+	public void handleContractNegotiationAccepted() {
+		String contractNegotaitionId = UUID.randomUUID().toString();
+		when(contractNegotiationRepository.findById(contractNegotaitionId))
+			.thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_OFFERED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(true);
+
+		ContractNegotiation response = service.handleContractNegotiationAccepted(contractNegotaitionId);
+		assertNotNull(response);
+		assertEquals(ContractNegotiationState.ACCEPTED, response.getState());
+	}
+	
+	@Test
+	@DisplayName("Consumer accepts contract negotiation offered by provider")
+	public void handleContractNegotiationAccepted_invalid_state() {
+		String contractNegotaitionId = UUID.randomUUID().toString();
+		when(contractNegotiationRepository.findById(contractNegotaitionId))
+			.thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
+
+		assertThrows(ContractNegotiationAPIException.class, 
+				()-> service.handleContractNegotiationAccepted(contractNegotaitionId));
+	}
+	
+	@Test
+	@DisplayName("Consumer accepts contract negotiation offered by provider - error api")
+	public void handleContractNegotiationAccepted_error_api() {
+		String contractNegotaitionId = UUID.randomUUID().toString();
+		when(contractNegotiationRepository.findById(contractNegotaitionId))
+			.thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_OFFERED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(false);
+
+		assertThrows(ContractNegotiationAPIException.class, 
+				()-> service.handleContractNegotiationAccepted(contractNegotaitionId));	
+		}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation")
+	public void handleCNApproved() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
+		when(properties.getAssignee()).thenReturn(ModelUtil.ASSIGNEE);
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.of(ModelUtil.OFFER));
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(true);
+		
+		service.handleContractNegotiationAgreed(contractNegotaitionId);
+		
+		verify(contractNegotiationRepository).save(argumentCaptor.capture());
+		assertEquals(ContractNegotiationState.AGREED, argumentCaptor.getValue().getState());
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - invalid initial state")
+	public void handleCNApproved_invalid_state() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_AGREED));
+		
+		assertThrows(ContractNegotiationAPIException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - offer does not exists")
+	public void handleCNApproved_no_offer() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.empty());
+		
+		assertThrows(OfferNotFoundException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+	}
+	
+	@Test
+	@DisplayName("Provider accepts contract negotiation - error while contacting consumer")
+	public void handleCNApproved_error_consumer() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(contractNegotiationRepository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(properties.providerCallbackAddress()).thenReturn(ModelUtil.CALLBACK_ADDRESS);
+		when(properties.getAssignee()).thenReturn(ModelUtil.ASSIGNEE);
+		when(offerRepository.findById(any(String.class))).thenReturn(Optional.of(ModelUtil.OFFER));
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(false);
+		
+		assertThrows(ContractNegotiationAPIException.class, 
+				() -> service.handleContractNegotiationAgreed(contractNegotaitionId));
+		
+		verify(contractNegotiationRepository, times(0)).save(argumentCaptor.capture());
+		verify(agreementRepository, times(0)).save(argCaptorAgreement.capture());
+	}
+
 }